fix: Page별 구조 변경
MyApp을 따로 만들고 MyApp에서 Navigation을 사용
